flis_django/requirements-dev.txt
-r requirements.txt
nose==1.1.2
unittest2==0.5.1
Fabric==1.4.3
ssh==1.7.14
pycrypto==2.6
mock==0.8.0
WebTest==2.0.9
six==1.4.1
WebOb==1.2.3
waitress==0.8.7
django-webtest==1.7.5
factory-boy==2.2.1
pyquery==1.2.6
lxml==3.2.4
cssselect==0.9.1
